backend/requirements.txt
django==3.0.3
djangorestframework==3.11.0
django-cors-headers==3.2.1
djangorestframework-simplejwt==4.4.0
